Conditionally enabling PDF support on macOs
OnOn macOs we rely on ghostscript being already installed on the system in order to enable PDF export.

Therefore we added a checker class (GhostScriptChecker.as) that looks for the "gs" executable and provides the path to it or null if not found. Based on this information, MAIDENS enables or not its "Export to PDF" menu item.